[DEV] Update ErrnoTest to not depend on implementation details
A good unit test must not depend on implementation details which
means if the internal implementation details of a tested module
change, that should not affect the unit test.

In the case of ErrnoTest, the returned string was expected to be
formatted in a certain way (NAME: strerror(errno)). One can decide
to add text to the returned string, replace ":" with something else,
reverse position of input parameters and so on. These are implementation
details.

What's really important when using Errno component is to see provided
informations (function name and errno converted into string) in the
returned string no matter how the text is formatted. Doing so, the developer
can change/improve its implementation without the fear to break something.