[DEV] Make it easy to integrate the service in a bigger project
One can download a .tar.gz of the last release and add additional
Makefiles to compile it (Android.mk, Android.bp, etc).

Someone else can upload the project as-is to a custom git server
and compile it using CMake or a custom build system.

In any cases, reducing the dependencies of the project might ease
its integration a lot: No other projects to download, build and
install first before being able to generate an executable.

Thus, CLI11 v1.9.0 and nlohmann/json v3.7.3 (header-only libraries)
have been installed into "third_party" folder making it no necessary
anymore in the build environment.

This is not a very good way of doing things. Indeed, if CLI11 or any
other dependencys is required, one should be aware of that and check
it out independently. However, putting everything necessary in the
tarball has its advantages.

Note: GTest/GMock still has to be made available if compiling tests
      is needed.